--- runbot.bat
@@ -1,12 +1,0 @@
-rem Set the OPENAI_API_KEY environment variable
-set OPENAI_API_KEY=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
-
-rem Set the DISCORD_BOT_TOKEN environment variable
-set DISCORD_BOT_TOKEN=MTA1MDkwNTIwMDc0NTAwMTAwMQ.GPqzO8.TI-VtQK5fgZTURYCQiV4DXQi-qR7SICgV0FVRU
-
-rem Set the DIALOGPT_AUTH_TOKEN environment variable
-set DIALOGPT_AUTH_TOKEN=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
-
-rem Run the Discord bot program using Python
-rem python -m pdb  botfaq.py
-python  botfaq.py
